geom_nodetext {ggnetwork}	R Documentation

Label the nodes of a network.

Description

All arguments to these geoms are identical to those of geom_text and geom_label.

Usage

geom_nodetext(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_nodelabel(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping
Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

position
A position adjustment to use on the data for this layer. Cannot be jointy specified with nudge_x or nudge_y. This can be used in various ways, including to prevent overplotting and improving the display. The position argument accepts the following:

The result of calling a position function, such as position_jitter().

A string nameing the position adjustment. To give the position as a string, strip the function name of the position_ prefix. For example, to use position_jitter(), give the position as "jitter".

For more information and other ways to specify the position, see the layer position documentation.

...
Other arguments passed on to layer()'s params argument. These arguments broadly fall into one of 4 categories below. Notably, further arguments to the position argument, or aesthetics that are required can not be passed through .... Unknown arguments that are not part of the 4 categories below are ignored.

Static aesthetics that are not mapped to a scale, but are at a fixed value and apply to the layer as a whole. For example, colour = "red" or linewidth = 3. The geom's documentation has an Aesthetics section that lists the available options. The 'required' aesthetics cannot be passed on to the params. Please note that while passing unmapped aesthetics as vectors is technically possible, the order and required length is not guaranteed to be parallel to the input data.

When constructing a layer using a stat_*() function, the ... argument can be used to pass on parameters to the geom part of the layer. An example of this is stat_density(geom = "area", outline.type = "both"). The geom's documentation lists which parameters it can accept.

Inversely, when constructing a layer using a geom_*() function, the ... argument can be used to pass on parameters to the stat part of the layer. An example of this is geom_area(stat = "density", adjust = 0.5). The stat's documentation lists which parameters it can accept.

The key_glyph argument of layer() may also be passed on through .... This can be one of the functions described as key glyphs, to change the display of the layer in the legend.

parse
If TRUE, the labels will be parsed into expressions and displayed as described in ?plotmath.

nudge_x, nudge_y
Horizontal and vertical adjustment to nudge labels by. Useful for offsetting text from points, particularly on discrete scales. Cannot be jointly specified with position.

check_overlap
If TRUE, text that overlaps previous text in the same layer will not be plotted. check_overlap happens at draw time and in the order of the data. Therefore data should be arranged by the label column before calling geom_text(). Note that this argument is not supported by geom_label().

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

label.padding
Amount of padding around label. Defaults to 0.25 lines.

label.r
Radius of rounded corners. Defaults to 0.15 lines.

label.size
Size of label border, in mm.

Examples

## geom_nodetext examples

if (require(network) && require(sna)) {
  n <- network(rgraph(10, tprob = 0.2), directed = FALSE)

  # just node labels
  ggplot(n, aes(x, y)) +
    geom_nodetext(aes(label = vertex.names)) +
    theme_blank()

  # with nodes underneath
  ggplot(n, aes(x, y)) +
    geom_nodes(colour = "gold", size = 9) +
    geom_nodetext(aes(label = vertex.names)) +
    theme_blank()

  # with nodes and edges
  ggplot(n, aes(x, y, xend = xend, yend = yend)) +
    geom_edges(colour = "gold") +
    geom_nodes(colour = "gold", size = 9) +
    geom_nodetext(aes(label = vertex.names)) +
    theme_blank()
}

## geom_nodelabel examples

if (require(network) && require(sna)) {
  data(flo, package = "network")
  n <- network(flo, directed = FALSE)

  # with text labels
  ggplot(n, aes(x, y, xend = xend, yend = yend)) +
    geom_edges(colour = "grey50") +
    geom_nodelabel(aes(label = vertex.names)) +
    theme_blank()

  # with text labels coloured according to degree centrality
  n %v% "degree" <- degree(n)
  ggplot(n, aes(x, y, xend = xend, yend = yend)) +
    geom_edges(colour = "grey50") +
    geom_nodelabel(aes(label = vertex.names, fill = degree)) +
    scale_fill_gradient(low = "gold", high = "tomato") +
    theme_blank()

  # label only a subset of all nodes
  high_degree <- function(x) {
    x[ x$degree > median(x$degree), ]
  }
  ggplot(n, aes(x, y, xend = xend, yend = yend)) +
    geom_edges(colour = "steelblue") +
    geom_nodes(aes(size = degree), colour = "steelblue") +
    geom_nodelabel(aes(label = vertex.names),
      data = high_degree,
      colour = "white", fill = "tomato"
    ) +
    theme_blank()
}

[Package ggnetwork version 0.5.14 Index]