ggnetwork {ggnetwork}	R Documentation

Fortify network objects.

Description

A wrapper for the fortify.network and fortify.igraph functions that will also try to coerce matrices and data frames to network objects.

Usage

ggnetwork(x, ...)

Arguments

x
an object of class network or igraph, or any object that can be coerced to that class, such as an adjacency or incidence matrix, or an edge list: see edgeset.constructors and network for details.

...
arguments passed to the fortify.network or fortify.igraph functions.

[Package ggnetwork version 0.5.14 Index]